requirements.txt
streamlit
openai
pandas
openpyxl
requests
Pillow